Stop using DB_PORT as WordPress code doesn't either
WP core only uses DB_HOST and it can have or not have the port as a suffix.
Removing it makes our config cleaner and also avoids confusing situations
where somebody changes the DB_PORT and is left wondering why it does not
have any effect (=because WP core does not use it, only DB_HOST).