Move PHP setting to separate file and bump it to 7.3
Latest WordPress 5.1 is supposed to work well with PHP 7.3 and it should
be a safe selection for new sites. Some plugins might not yet be fully
PHP 7.3 compatible, but since this applies to new sites it is a safe thing
to do and can even lead to customers reporting PHP 7.3 issues to us so we
get more feedback on the general PHP 7.3 compatibility status.