Replace vagrant triggers plugin with core triggers (Closes: #88)
The idea is to not force developers to immediately update their Vagrant
versions, and provide a smooth backwards compatibility to the usage of the
vagrant-triggers plugin. The Vagrantfile now contains definitions for both
Vagrant core and vagrant-triggers plugin triggers, and the decision between
the two is made with the following logic:
- if the user uses a Vagrant version with support for core triggers and
vagrant-triggers is not installed, then use core triggers directly
- if the user uses a Vagrant version with support for core triggers but still
has the vagrant-triggers plugin installed, he/she can either uninstall the
plugin or define VAGRANT_USE_VAGRANT_TRIGGERS as an environment variable so
that Vagrant internally knows to use the vagrant-triggers plugin syntax
- if an older version of Vagrant is used, then the vagrant-triggers plugin
is required to be installed

After this we can advise our customers to start updating their Virtualbox
and Vagrant versions (and uninstall vagrant-triggers). After that, we can
proceed to drop the backwards-compatibility support for the vagrant-triggers
plugin somewhere in the future.