Replace old default plugins with new reliable ones from Google
We've recommended auto-post-thumbnail and google-analytics-dashboard-for-wp.
However both of them seem to have been thwarted lately into sucipicious use
so it will be better to ship by default plugins that have an reliable
author and is unlikely to be misused for other purposes than purely server
the WordPress site.

Google Site Kit replaces the functionality of google-analytics-dashboard-for-wp.

The Native Lazyload plugin does not replace the auto-post-thumbnail plugin,
but is a pretty safe speed optimization, so it's a good one to promote.

Closes: #104
Closes: #110