Rely on running wp-development-up instead of custom Vagrantfile code
The code in the Vagrantfile is bloated and obsolete anyway, it should
simply be running wp-development-up after the Vagrant box is up so
that all imports and pulls etc run correctly inside the Vagrant box without
depending on the Vagrantfile contents. This makes the project template
also better compatible and unified with using Docker.

Related: #164.

Also clean away some extra output to avoid excess verbosity.