Add the .nvmrc file to facilitate the use of nvm
With .nvmrc the Node Version Manager (nvm) can be used to manage
the NodeJS version dependency. Dependencies introduced in the
package.json file have an implicit restriction on which NodeJS
version(s) they can be used. The .nvmrc file makes this restriction
explicit.

After installing nvm you can make use of this file by executing the
following commands:

```
$ nvm install
$ nvm use
```

This installs and activates the correct NodeJS version for the
project. After this `npm install` and `gulp` should work without
a hitch.

Co-authored-by: Kimmo Tapala <kimitri@gmail.com>